feat: 사용자 프로필 페이지에 설정 및 정보 수정 버튼 추가
사용자 프로필 페이지에 설정 버튼을 추가하여 사용자가 설정을 쉽게 접근할 수 있도록 하였으며, 정보 수정 버튼을 통해 사용자 정보 편집 기능을 구현하였습니다. 이를 통해 사용자 경험을 향상시키고, 프로필 관리 기능을 더욱 직관적으로 개선하였습니다.
